Fix PluginContainer. Replaces boost::fusion::map with std::tuple.
PluginMap and m_Plugins have been added back as std::tuples

PluginContainer::plugins() has been adapted to tuples and returns a reference again. Also provided a const and non const overload.

All uses in the source file of bf::at_key<> have been replaced with this->plugins<>, which is the same thing and idk why it didnt do that in the first place.

Replaced some qPrintable with the correct qUtf8Printable

The only remaining hurdle for this class is the signals usage, but other than that it should be functionally equalivant