Start documenting SelectionDialog class and remove duplicate code.
The Duplicate QIcon addChoice overload is trivally able to be handled with a default paramater since QIcon default constructs to a null icon, and provides a method to check if it is null.

If for some reason a null icon is desired, a optional bool paramater to force it is trivial.