Boot process and Argument handling work.
Further work to consolidate argument handling, the nxm handling for the primary instance has been moved into the handleArguments function.
The new behaviour is to start a secondary instance with the nxm link.
The timeout for connecting to the socket in MoIPC::sendMessage has also been changed to the default 30 seconds, to give more time in this scenario.

handleArguments has been changed to return a bool.
True indicates the program should terminate after handling arguments, and false indicates it should continue starting up.

The update argument has been removed.