Remove redundant m_ValidateByData variable.
This variable doesnt seem to do anything, and is used incorrectly.

m_ValidateByData will always be true if m_Choice->property("data").isValid(), so a condition checking if one is true and the other is false will always be false.

Likewise, a condition checking if it is false OR if the other is true will always be true.